tests/wpt/css-tests/compositing-1_dev/xhtml1print/mix-blend-mode-parsing.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <meta charset="utf-8" />
        <title>CSS Test: parsing mix-blend-mode property</title>
        <link href="mailto:mbudaes@adobe.com" rel="author" title="Mirela Budăeș" />
        <link href="mailto:mibalan@adobe.com" rel="reviewer" title="Mihai Balan" />
        <link href="mailto:cabanier@adobe.com" rel="reviewer" title="Rik Cabanier" />
        <link href="https://drafts.fxtf.org/compositing-1/#mix-blend-mode" rel="help" />
        <meta content="dom" name="flags" />
        <meta content="Test checks that the value specified for mix-blend-mode property is correctly parsed" name="assert" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <link href="/resources/testharness.css" rel="stylesheet" />
    </head>
    <body>
        <div id="log"></div>
        <script type="text/javascript">
            function testParse(declaration) {
                var div = document.createElement("div");
                div.setAttribute("style", declaration);
                return div.style.mixBlendMode;
            }

            var blendModes = ["normal", "multiply", "screen", "overlay", "darken", "lighten", "color-dodge", "color-burn",
                      "hard-light", "soft-light", "difference", "exclusion", "hue", "saturation", "color", "luminosity"];

            for (var i = 0; i &lt; blendModes.length; i++) {
                test(function() {assert_equals(testParse("mix-blend-mode: " + blendModes[i]), blendModes[i] )}, "Mix-blend-mode " + blendModes[i]);
            };
        </script>
    

</body></html>
